Рефакторинг работы с шаблоном таймера - создан общий вью уровня
