+++ js/views/abstract-view.js
@@ -7,11 +7,6 @@ export default class AbstractView {
     }
   }
 
-  set timer(time) {
-    this._timer.minutesNode.innerText = time.normalizedMinutes;
-    this._timer.secondsNode.innerText = time.normalizedSeconds;
-  }
-
   get template() {
     throw new Error(`Template is required`);
   }
