+++ js/views/level-genre-view.js
@@ -1,7 +1,6 @@
-import AbstractView from './abstract-view';
-import timerTemplate from './timer-template';
+import LevelView from './level-view';
 
-export default class LevelGenreView extends AbstractView {
+export default class LevelGenreView extends LevelView {
   constructor(level, time) {
     super();
     this.level = level;
@@ -15,7 +14,7 @@ export default class LevelGenreView extends AbstractView {
   get template() {
     return `
     <section class="main main--level main--level-genre">
-      ${timerTemplate(this.time)}
+      ${this.timerTemplate}
       <div class="main-mistakes">
         ${`<img class="main-mistake" src="img/wrong-answer.png" width="35" height="49">`.repeat(this.mistakes())}
       </div>
